refactor(gui): move filter model classes into filter feature
