+++ api-editor/gui/src/app/App.tsx
@@ -65,7 +65,7 @@ import { OptionalBatchForm } from '../features/annotations/batchforms/OptionalBa
 import { RemoveBatchForm } from '../features/annotations/batchforms/RemoveBatchForm';
 import { MoveBatchForm } from '../features/annotations/batchforms/MoveBatchForm';
 import { PythonPackage } from '../features/packageData/model/PythonPackage';
-import { AbstractPythonFilter } from '../features/packageData/model/filters/AbstractPythonFilter';
+import { AbstractPythonFilter } from '../features/filter/model/AbstractPythonFilter';
 import { UsageCountStore } from '../features/usages/model/UsageCountStore';
 import { PythonDeclaration } from '../features/packageData/model/PythonDeclaration';
 import { SaveFilterDialog } from '../features/filter/SaveFilterDialog';
