+++ api-editor/gui/src/features/filter/model/NameStringFilter.ts
@@ -1,25 +1,22 @@
-import { PythonClass } from '../PythonClass';
-import { PythonFunction } from '../PythonFunction';
-import { PythonModule } from '../PythonModule';
-import { PythonParameter } from '../PythonParameter';
+import { PythonClass } from '../../packageData/model/PythonClass';
+import { PythonFunction } from '../../packageData/model/PythonFunction';
+import { PythonModule } from '../../packageData/model/PythonModule';
+import { PythonParameter } from '../../packageData/model/PythonParameter';
+import { PythonDeclaration } from '../../packageData/model/PythonDeclaration';
+import { AnnotationStore } from '../../annotations/annotationSlice';
+import { UsageCountStore } from '../../usages/model/UsageCountStore';
 import { AbstractPythonFilter } from './AbstractPythonFilter';
-import { PythonDeclaration } from '../PythonDeclaration';
-import { AnnotationStore } from '../../../annotations/annotationSlice';
-import { UsageCountStore } from '../../../usages/model/UsageCountStore';
 
 /**
- * Keeps only declarations that have a name matching the given regex.
+ * Keeps only declarations that have a given string in their name.
  */
-export class NameRegexFilter extends AbstractPythonFilter {
-    readonly regex: RegExp;
-
+export class NameStringFilter extends AbstractPythonFilter {
     /**
-     * @param regex The regex that must match the name of the declaration.
+     * @param string The string that must be part of the name of the declaration.
+     * @param matchExactly Whether the name must match the substring exactly.
      */
-    constructor(regex: string) {
+    constructor(readonly string: string, readonly matchExactly: boolean) {
         super();
-
-        this.regex = RegExp(regex, 'u');
     }
 
     shouldKeepModule(pythonModule: PythonModule, annotations: AnnotationStore, usages: UsageCountStore): boolean {
@@ -47,6 +44,10 @@ export class NameRegexFilter extends AbstractPythonFilter {
         _annotations: AnnotationStore,
         _usages: UsageCountStore,
     ): boolean {
-        return this.regex.test(pythonDeclaration.name);
+        if (this.matchExactly) {
+            return pythonDeclaration.name === this.string;
+        } else {
+            return pythonDeclaration.name.includes(this.string);
+        }
     }
 }
