+++ api-editor/gui/src/features/filter/model/VisibilityFilter.ts
@@ -1,21 +1,20 @@
-import { PythonClass } from '../PythonClass';
-import { PythonFunction } from '../PythonFunction';
-import { PythonModule } from '../PythonModule';
-import { PythonParameter } from '../PythonParameter';
+import { PythonClass } from '../../packageData/model/PythonClass';
+import { PythonFunction } from '../../packageData/model/PythonFunction';
+import { PythonModule } from '../../packageData/model/PythonModule';
+import { PythonParameter } from '../../packageData/model/PythonParameter';
+import { PythonDeclaration } from '../../packageData/model/PythonDeclaration';
+import { AnnotationStore } from '../../annotations/annotationSlice';
+import { UsageCountStore } from '../../usages/model/UsageCountStore';
 import { AbstractPythonFilter } from './AbstractPythonFilter';
-import { PythonDeclaration } from '../PythonDeclaration';
-import { AnnotationStore } from '../../../annotations/annotationSlice';
-import { UsageCountStore } from '../../../usages/model/UsageCountStore';
 
 /**
- * Keeps only declarations that have a given string in their name.
+ * Keeps only declarations with a specified visibility (public/internal)
  */
-export class NameStringFilter extends AbstractPythonFilter {
+export class VisibilityFilter extends AbstractPythonFilter {
     /**
-     * @param string The string that must be part of the name of the declaration.
-     * @param matchExactly Whether the name must match the substring exactly.
+     * @param visibility The visibility of the declaration to keep.
      */
-    constructor(readonly string: string, readonly matchExactly: boolean) {
+    constructor(readonly visibility: Visibility) {
         super();
     }
 
@@ -44,10 +43,11 @@ export class NameStringFilter extends AbstractPythonFilter {
         _annotations: AnnotationStore,
         _usages: UsageCountStore,
     ): boolean {
-        if (this.matchExactly) {
-            return pythonDeclaration.name === this.string;
-        } else {
-            return pythonDeclaration.name.includes(this.string);
-        }
+        return pythonDeclaration.isPublic === (this.visibility === Visibility.Public);
     }
+}
+
+export enum Visibility {
+    Public,
+    Internal,
 }
