+++ api-editor/gui/src/features/filter/model/comparisons.ts
