+++ api-editor/gui/src/features/packageData/treeView/FunctionNode.tsx
@@ -3,7 +3,7 @@ import { FaCogs } from 'react-icons/fa';
 import { isEmptyList } from '../../../common/util/listOperations';
 import { PythonFunction } from '../model/PythonFunction';
 import { TreeNode, ValuePair } from './TreeNode';
-import { AbstractPythonFilter } from '../model/filters/AbstractPythonFilter';
+import { AbstractPythonFilter } from '../../filter/model/AbstractPythonFilter';
 import { UsageCountStore } from '../../usages/model/UsageCountStore';
 import { useAppSelector } from '../../../app/hooks';
 import { maximumNumberOfFunctionAnnotations, selectNumberOfAnnotations } from '../../annotations/annotationSlice';
